Fixes #20776: Address PR review feedback
Restore BulkRenameForm to utilities.forms (preserving the existing
import path) and introduce NetBoxModelBulkRenameForm in netbox.forms
as a changelog-aware wrapper, following the established
NetBoxModelBulkEditForm pattern.

BulkRenameView now conditionally selects the base form based on
whether the model supports change logging, eliminating the need
to mutate base_fields after class creation.
