+++ netbox/utilities/forms/forms.py
@@ -1,3 +1,5 @@
+import re
+
 from django import forms
 from django.utils.translation import gettext as _
 
@@ -8,6 +10,7 @@ from utilities.forms.mixins import BackgroundJobMixin, FilterModifierMixin
 __all__ = (
     'BulkDeleteForm',
     'BulkEditForm',
+    'BulkRenameForm',
     'CSVModelForm',
     'ConfirmationForm',
     'DeleteForm',
@@ -56,6 +59,36 @@ class BulkEditForm(BackgroundJobMixin, forms.Form):
         nullable_fields: A list of field names indicating which fields support being set to null/empty
     """
     nullable_fields = ()
+
+
+class BulkRenameForm(forms.Form):
+    """
+    An extendable form to be used for renaming objects in bulk.
+    """
+    find = forms.CharField(
+        strip=False
+    )
+    replace = forms.CharField(
+        strip=False,
+        required=False
+    )
+    use_regex = forms.BooleanField(
+        required=False,
+        initial=True,
+        label=_('Use regular expressions')
+    )
+
+    def clean(self):
+        super().clean()
+
+        # Validate regular expression in "find" field
+        if self.cleaned_data['use_regex']:
+            try:
+                re.compile(self.cleaned_data['find'])
+            except re.error:
+                raise forms.ValidationError({
+                    'find': "Invalid regular expression"
+                })
 
 
 class BulkDeleteForm(BackgroundJobMixin, ConfirmationForm):
